test: AcceptanceTest 추상화
