+++ src/main/java/com/portfolio/rebalancer/domain/color/Color.java
@@ -16,6 +16,7 @@ public class Color {
 
 	public static final String REGEX = "^#([a-fA-F0-9]{6})$";
 	static final String REGEX_ERROR_MESSAGE = "색상 형식이 맞지 않습니다.";
+
 	@Column(name = "color", nullable = false)
 	private String value;
