+++ src/test/java/com/portfolio/rebalancer/acceptance/AcceptanceTest.java
@@ -1,0 +1,37 @@
+package com.portfolio.rebalancer.acceptance;
+
+import org.junit.jupiter.api.BeforeEach;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.web.server.LocalServerPort;
+import org.springframework.http.MediaType;
+
+import com.portfolio.rebalancer.support.DatabaseCleanUp;
+
+import io.restassured.RestAssured;
+import io.restassured.response.ValidatableResponse;
+
+@SpringBootTest(properties = "spring.session.store-type=none", webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+public class AcceptanceTest {
+
+	@LocalServerPort
+	int port;
+
+	@Autowired
+	private DatabaseCleanUp databaseCleanUp;
+
+	@BeforeEach
+	void setUp() {
+		RestAssured.port = port;
+		databaseCleanUp.execute();
+	}
+
+	protected ValidatableResponse post(String uri, Object request) {
+		return RestAssured.given().log().all()
+			.body(request)
+			.contentType(MediaType.APPLICATION_JSON_VALUE)
+			.accept(MediaType.APPLICATION_JSON_VALUE)
+			.when().post(uri)
+			.then().log().all();
+	}
+}
